backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiohttp==3.10.5
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2
certifi==2024.8.30
